Move utils/dom-fingerprint to core/dom/fingerprint
