+++ src/core/dom/fingerprint.js
