+++ test/unit/core/dom/test-fingerprint.js
@@ -17,7 +17,7 @@
 import {
   DomFingerprint,
   domFingerprintPlain,
-} from '../../../src/utils/dom-fingerprint';
+} from '../../../../src/core/dom/fingerprint';
 
 describes.realWin('domFingerprint', {}, (env) => {
   let body;
